+++ src/pages/Dashboard/index.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import React, { useContext } from 'react';
 import { AiOutlineUser } from 'react-icons/ai';
